backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
httpx==0.27.2
# Align all LangChain packages on core 1.x to avoid resolver backtracking
langchain-core>=1.2,<2
langchain-openai>=1.0.0
langchain-anthropic>=1.0.0
langchain-deepseek>=1.0.0
jsonschema==4.23.0
langchain-mcp-adapters
